+++ src/WebAppService.php
@@ -2,14 +2,15 @@
 
 namespace Bayfront\BonesService\WebApp;
 
-use Bayfront\ArrayHelpers\Arr;
 use Bayfront\Bones\Abstracts\Service;
 use Bayfront\Bones\Application\Services\Events\EventService;
 use Bayfront\Bones\Application\Services\Filters\FilterService;
+use Bayfront\Bones\Application\Utilities\App;
 use Bayfront\Bones\Exceptions\ServiceException;
 use Bayfront\BonesService\WebApp\Events\WebAppServiceEvents;
 use Bayfront\BonesService\WebApp\Exceptions\WebAppServiceException;
 use Bayfront\BonesService\WebApp\Filters\WebAppServiceFilters;
+use Bayfront\Container\NotFoundException;
 use Bayfront\HttpResponse\InvalidStatusCodeException;
 use Bayfront\HttpResponse\Response;
 use Bayfront\RouteIt\Router;
@@ -24,7 +25,6 @@ class WebAppService extends Service
     public Response $response;
     public Router $router;
     public Veil $veil;
-    public array $config;
 
     /**
      * The container will resolve any dependencies.
@@ -35,51 +35,42 @@ class WebAppService extends Service
      * @param Response $response
      * @param Router $router
      * @param Veil $veil
-     * @param array $config
      * @throws WebAppServiceException
      */
 
-    public function __construct(EventService $events, FilterService $filters, Response $response, Router $router, Veil $veil, array $config)
+    public function __construct(EventService $events, FilterService $filters, Response $response, Router $router, Veil $veil)
     {
         $this->events = $events;
         $this->filters = $filters;
         $this->response = $response;
         $this->router = $router;
         $this->veil = $veil;
-        $this->config = $config;
 
         parent::__construct($events);
+
+        if (!App::has('Bayfront\Translation\Translate')) {
+            throw new WebAppServiceException('Unable to start WebAppService: Required dependency not found (Bayfront\Translation\Translate)');
+        }
 
         // Enqueue events
 
         try {
-            $this->events->addSubscriptions(new WebAppServiceEvents($this));
-        } catch (ServiceException $e) {
+            $translate = App::get('Bayfront\Translation\Translate');
+            $this->events->addSubscriptions(new WebAppServiceEvents($this, $translate));
+        } catch (ServiceException|NotFoundException $e) {
             throw new WebAppServiceException('Unable to start WebAppService: ' . $e->getMessage(), $e->getCode(), $e->getPrevious());
         }
 
         // Enqueue filters
 
         try {
-            $this->filters->addSubscriptions(new WebAppServiceFilters($this));
+            $this->filters->addSubscriptions(new WebAppServiceFilters($this, $translate));
         } catch (ServiceException $e) {
             throw new WebAppServiceException('Unable to start WebAppService: ' . $e->getMessage(), $e->getCode(), $e->getPrevious());
         }
 
         $this->events->doEvent('webapp.start', $this);
 
-    }
-
-    /**
-     * Get web app configuration value in dot notation.
-     *
-     * @param string $key (Key to return in dot notation)
-     * @param mixed|null $default (Default value to return if not existing)
-     * @return mixed
-     */
-    public function getConfig(string $key = '', mixed $default = null): mixed
-    {
-        return Arr::get($this->config, $key, $default);
     }
 
     /**