+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `WebAppService` config array to Veil data array
+- Added locale functionality
+- Added `@route` and `@say` template tags
 
 ## [1.2.0] - 2025.01.09